refactor folder structure add test
