+++ js/packages/openinference-core/src/utils/typeUtils.ts
@@ -1,5 +1,10 @@
 import { AttributeValue } from "@opentelemetry/api";
 
+/**
+ * Type guard to determine whether or not a value is an array of nullable numbers.
+ * @param value
+ * @returns true if the value is an array of nullable numbers, false otherwise.
+ */
 function isSparseNumericArray(
   value: unknown,
 ): value is (number | null | undefined)[] {
@@ -9,6 +14,11 @@ function isSparseNumericArray(
   );
 }
 
+/**
+ * Type guard to determine whether or not a value is an array of nullable booleans.
+ * @param value
+ * @returns true if the value is an array of nullable booleans, false otherwise.
+ */
 function isSparseBooleanArray(
   value: unknown,
 ): value is (boolean | null | undefined)[] {
@@ -18,6 +28,11 @@ function isSparseBooleanArray(
   );
 }
 
+/**
+ * Type guard to determine whether or not a value is an array of nullable strings.
+ * @param value
+ * @returns true if the value is an array of nullable strings, false otherwise.
+ */
 function isSparseStringArray(
   value: unknown,
 ): value is (string | null | undefined)[] {
@@ -27,6 +42,11 @@ function isSparseStringArray(
   );
 }
 
+/**
+ * Type guard to determine whether or not a value is an attribute value (i.e., a primitive value or an array of strings, numbers or booleans).
+ * @param value
+ * @returns true if the value is an attribute value, false otherwise.
+ */
 export function isAttributeValue(value: unknown): value is AttributeValue {
   return (
     typeof value === "string" ||
